Refactor some tests to be more specific
